+++ frontend/app/vite.config.ts
@@ -1,11 +1,56 @@
+import path from "node:path";
+import tailwindcss from "@tailwindcss/vite";
+import { TanStackRouterVite } from "@tanstack/router-plugin/vite";
 import react from "@vitejs/plugin-react";
+import type { Plugin, PluginOption } from "vite";
 import { defineConfig } from "vite";
+import svgr from "vite-plugin-svgr";
+
+const plugins: (Plugin<unknown> | PluginOption[])[] = [
+  tailwindcss(),
+  TanStackRouterVite({
+    target: "react",
+    autoCodeSplitting: true,
+    routesDirectory: "../core/src/routes",
+    generatedRouteTree: "../core/src/routeTree.gen.ts",
+  }),
+  react({
+    babel: {
+      presets: ["jotai/babel/preset"],
+    },
+  }),
+  svgr({
+    svgrOptions: {
+      exportType: "named",
+      ref: true,
+    },
+  }),
+];
+
+const coreRoot = path.resolve(__dirname, "../core");
+const alias = {
+  "@": path.resolve(coreRoot, "src"),
+  "@lib": path.resolve(coreRoot, "src/scripts/lib"),
+  "@components": path.resolve(coreRoot, "src/scripts/components"),
+  "@pages": path.resolve(coreRoot, "src/scripts/pages"),
+  "@public": path.resolve(coreRoot, "public"),
+  "@hooks": path.resolve(coreRoot, "src/scripts/hooks"),
+};
 
 const host = process.env.TAURI_DEV_HOST;
 
-// https://vitejs.dev/config/
-export default defineConfig(async () => ({
-  plugins: [react()],
+export default defineConfig(() => ({
+  plugins,
+  optimizeDeps: {
+    exclude: ["@decopon/core"],
+  },
+  publicDir: "../core/public",
+  resolve: {
+    alias: {
+      ...alias,
+      "@decopon/core": path.resolve(__dirname, "../core/src"),
+    },
+  },
 
   // Vite options tailored for Tauri development and only applied in `tauri dev` or `tauri build`
   //
