+++ frontend/core/buildSettings/alias.ts
@@ -1,19 +1,21 @@
-import path from "node:path";
 // import path from "node:path";
-import { fileURLToPath } from "node:url";
+// // import path from "node:path";
+// import { fileURLToPath } from "node:url";
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+// const __filename = fileURLToPath(import.meta.url);
+// const __dirname = path.dirname(__filename);
 
-const root = path.resolve(__dirname, "..");
+// const root = path.resolve(__dirname, "..");
 
-const alias = {
-  "@": path.resolve(root, "src"),
-  "@lib": path.resolve(root, "src/scripts/lib"),
-  "@components": path.resolve(root, "src/scripts/components"),
-  "@pages": path.resolve(root, "src/scripts/pages"),
-  "@public": path.resolve(root, "public"),
-  "@hooks": path.resolve(root, "src/scripts/hooks"),
-};
+// const alias = {
+//   "@": path.resolve(root, "src"),
+//   "@lib": path.resolve(root, "src/scripts/lib"),
+//   "@components": path.resolve(root, "src/scripts/components"),
+//   "@pages": path.resolve(root, "src/scripts/pages"),
+//   "@public": path.resolve(root, "public"),
+//   "@hooks": path.resolve(root, "src/scripts/hooks"),
+// };
 
-export default alias;
+// console.log("Alias configuration:", alias);
+
+// export default alias;
