+++ frontend/core/src/index.ts
@@ -1,8 +1,4 @@
-export { App } from "@/scripts/App";
-export { Ziggy } from "@lib/ziggy.js";
+export { App } from "./scripts/App";
+export { Ziggy } from "./scripts/lib/ziggy.js";
 export { route } from "ziggy-js";
-export type { Ziggy as ZiggyType, route as ziggyRoute } from "@lib/ziggy.js";
-
-// build settings
-export { default as plugins } from "../buildSettings/plugins";
-export { default as alias } from "../buildSettings/alias";
+export type { Ziggy as ZiggyType } from "./scripts/lib/ziggy.js";
